requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.2
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
charset-normalizer==3.4.0
Eel==0.17.0
future==1.0.0
gevent==24.10.2
gevent-websocket==0.10.1
greenlet==3.1.1
idna==3.10
macholib==1.16.3
mysql-connector-python==9.0.0
packaging==24.1
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.1.0
